+++ Source/StageBuild.swift
@@ -3,4 +3,4 @@ import Foundation
 /**
  Tuple containing the stage and its minimum build number.
  */
-public typealias StageBuild = (stage: Stage, minimum: UInt8)
+public typealias StageBuild = (stage: Stage, minimum: Int)
